Fix definition of templated objects
Remove the cpp files from CMake's source files and instead include the
cpp files in their respective headers. To the compiler, it will look
like the definitions were written in the same file as the headers,
albeit not inline (i.e. outside the class declaration). This is kind of
hacky, but it allows the definitions to be visually separate from the
declarations.