Fix DFS not returning shortest path due to order of neighbours
Since vertices were always marked as visited, it prevented the search
from later revisiting those nodes when it was trying to go down a
shorter path. This meant the algorithm only returned the shortest path
if the vertices leading to that path happened to be listed before
other neighbours that were part of longer paths.