Avoid rebuilds in repl.it
Originally, the CMake build directory was in the repl.it runner's home
directory. This meant that if the author of the repl.it ran the program,
it would write and persist the cmake build directory. However, CMake
generates files which contain absolute references to the runner's home
directory. The directory's name is different for guest viewers, meaning
CMake would fail to build.

The solution was to delete and re-create the build directory upon every
run. Instead, move the build directory out of the runner's home to avoid
persisting it.